+++ src/Covenant/Internal/Term.hs
@@ -149,6 +149,10 @@ data CovenantTypeError
     --
     -- @since 1.1.0
     CataUnsuitable (CompT AbstractTy) (ValT AbstractTy)
+  | -- | Someone attempted to construct a tyvar using a DB index or argument position
+    --   which refers to a scope (or argument) that does not exist.
+    -- @since 1.2.0
+    OutOfScopeTyVar DeBruijn (Index "tyvar")
   deriving stock
     ( -- | @since 1.0.0
       Eq,